+++ kiwixbuild/dependencies/icu4c.py
@@ -28,7 +28,7 @@ class Icu(Dependency):
         @property
         def configure_option(self):
             options = "--disable-samples --disable-tests --disable-extras --disable-dyload --enable-rpath"
-            if self.buildEnv.platform_info.build == 'android':
+            if self.buildEnv.platformInfo.build == 'android':
                 options += " --with-data-packaging=archive"
             return options
