+++ kiwixbuild/platforms/native.py
@@ -5,9 +5,6 @@ class NativePlatformInfo(PlatformInfo):
     def __init__(self, name, static, hosts):
         super().__init__(name, 'native', static, [], hosts)
 
-    def get_cross_config(self):
-        return {}
-
 
 NativePlatformInfo('native_dyn', False, ['fedora', 'debian', 'Darwin'])
 NativePlatformInfo('native_static', True, ['fedora', 'debian'])
