+++ kiwixbuild/templates/cmake_android_cross_file.txt
@@ -1,7 +1,8 @@
 SET(CMAKE_SYSTEM_NAME {host_machine[system]})
-SET(CMAKE_ANDROID_STANDALONE_TOOLCHAIN {toolchain.builder.install_path})
 
-SET(CMAKE_C_COMPILER "{toolchain.binaries[CC]}")
-SET(CMAKE_CXX_COMPILER "{toolchain.binaries[CXX]}")
+SET(CMAKE_ANDROID_STANDALONE_TOOLCHAIN {install_path})
 
-SET(CMAKE_FIND_ROOT_PATH {toolchain.root_path})
+SET(CMAKE_C_COMPILER "{binaries[CC]}")
+SET(CMAKE_CXX_COMPILER "{binaries[CXX]}")
+
+SET(CMAKE_FIND_ROOT_PATH {root_path})
