+++ kiwixbuild/templates/cmake_i586_cross_file.txt
@@ -2,13 +2,13 @@ SET(CMAKE_SYSTEM_NAME {host_machine[system]})
 SET(CMAKE_SYSTEM_PROCESSOR {host_machine[cpu_family]})
 
 # specify the cross compiler
-SET(CMAKE_C_COMPILER "{toolchain.binaries[CC]}")
-SET(CMAKE_CXX_COMPILER "{toolchain.binaries[CXX]}")
+SET(CMAKE_C_COMPILER "{binaries[CC]}")
+SET(CMAKE_CXX_COMPILER "{binaries[CXX]}")
 SET(C_FLAGS "-m32 -march=i586")
 SET(CXX_FLAGS "-m32 -march=i586")
 SET(CMAKE_LD_FLAGS "-m32 -march=i586")
-SET(CMAKE_AR:FILEPATH {toolchain.binaries[AR]})
-SET(CMAKE_RANLIB:FILEPATH {toolchain.binaries[RANLIB]})
+SET(CMAKE_AR:FILEPATH {binaries[AR]})
+SET(CMAKE_RANLIB:FILEPATH {binaries[RANLIB]})
 
 find_program(CCACHE_FOUND ccache)
 if(CCACHE_FOUND)
