+++ kiwixbuild/templates/cmake_ios_cross_file.txt
@@ -1,7 +1,7 @@
 SET(CMAKE_SYSTEM_NAME {host_machine[system]})
 
-SET(CMAKE_C_COMPILER "{toolchain.binaries[CC]}")
-SET(CMAKE_CXX_COMPILER "{toolchain.binaries[CXX]}")
+SET(CMAKE_C_COMPILER "{binaries[CC]}")
+SET(CMAKE_CXX_COMPILER "{binaries[CXX]}")
 
 SET(CMAKE_FIND_ROOT_PATH {root_path})
