+++ kiwixbuild/toolchains/__init__.py
@@ -1,4 +1,4 @@
 
-from . import android_ndk, android_sdk, armhf, ios, mingw32, linux_i586
+from . import android_ndk, android_sdk, armhf
 
 from .base_toolchain import Toolchain
