+++ kiwixbuild/toolchains/android_sdk.py
@@ -49,6 +49,3 @@ class android_sdk(Toolchain):
         def build(self):
             self.command('build_platform', self._build_platform)
             self.command('fix_licenses', self._fix_licenses)
-
-    def set_env(self, env):
-        env['ANDROID_HOME'] = self.builder.install_path
